solvers: move checksol() to utils submodule
